requirements.txt
asgiref==3.11.1
Django==5.2.11
django-unfold==0.78.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
